Break up the protocol.js file into individual files. (#81)
