+++ packages/core/src/datetimeUtils.ts
@@ -1,0 +1,3 @@
+export function dateToUnixSeconds(date: Date) {
+  return Math.floor(date.getTime() / 1000);
+}
