+++ packages/core/src/index.ts
@@ -1,8 +1,15 @@
-export * from "./Currency.js";
-export * from "./KycStatus.js";
 export * from "./NonceValidator.js";
-export * from "./PayerData.js";
-export * from "./protocol.js";
+export * from "./protocol/CounterPartyData.js";
+export * from "./protocol/Currency.js";
+export * from "./protocol/KycStatus.js";
+export * from "./protocol/LnurlpRequest.js";
+export * from "./protocol/LnurlpResponse.js";
+export * from "./protocol/PayeeData.js";
+export * from "./protocol/PayerData.js";
+export * from "./protocol/PayReqResponse.js";
+export * from "./protocol/PayRequest.js";
+export * from "./protocol/PostTransactionCallback.js";
+export * from "./protocol/PubKeyResponse.js";
 export * from "./PublicKeyCache.js";
 export * from "./uma.js";
 export * from "./UmaInvoiceCreator.js";
