+++ packages/core/src/protocol/CounterPartyData.ts
@@ -1,0 +1,21 @@
+import { z } from "zod";
+
+export const CounterPartyDataOptionSchema = z.object({
+  mandatory: z.boolean(),
+});
+
+export type CounterPartyDataOption = z.infer<
+  typeof CounterPartyDataOptionSchema
+>;
+
+/**
+ * CounterPartyDataOptions describes which fields a vasp needs to know about the sender or receiver.
+ * Used for payerData and payeeData.
+ */
+export const CounterPartyDataOptionsSchema = z.record(
+  CounterPartyDataOptionSchema,
+);
+
+export type CounterPartyDataOptions = z.infer<
+  typeof CounterPartyDataOptionsSchema
+>;
